update decoders, handle interface breakages
yaxpeax_arch abstracts decoders a little more; `yaxpeax_arch::Decoder`
now takes a `yaxpeax_arch::Reader` (rather than iterator of u8). this is
to prepare for architectures where the instruction's data primitive is
*not* a u8, but an u4, u2, u1, or some other value. some decoders
(msp43) now in fact use a word size of U16le, meaning yaxpeax_core has
to become flexible over varied word sizes already.

since instruction word sizes are attached to an Arch impl, the resulting
change was that `MemoryRepr` (and related types) should be parameterized
on `Arch`, instead of just the associated `Address` type.

in the process, an impl `MemoryRepr<A> for [u8]` became clear, and now
slices can be used with `yaxpeax_core` directly. some high level
interfaces may still assume `Vec<u8>`, and likely can be fixed.

+++ src/memory/reader/hex.rs
@@ -1,7 +1,6 @@
 use std::collections::HashMap;
 
 use memory::repr::FlatMemoryRepr;
-use memory::PatchyMemoryRepr;
 
 fn read_byte<'a, T: Iterator<Item=&'a u8>>(data: T) -> Result<u8, String> {
     let hex = data.take(2).map(|x| *x).collect::<Vec<u8>>();
@@ -72,7 +71,7 @@ pub fn from_hex(data: &Vec<u8>) -> Result<HashMap<u8, FlatMemoryRepr>, String> {
                     0 => {
                         // data
                         let mem = repr.entry(section).or_insert_with(|| FlatMemoryRepr::empty("from_hex".to_string()));
-                        mem.add(data, address).unwrap();
+                        mem.add(data, address as usize).unwrap();
                         Ok(())
                     },
                     1 => {
